wagtail_site/pyproject.toml
[tool.poetry]
name = "ssn-wagtail"
version = "0.1.0"
description = ""
authors = ["Chapin Bryce <mail@chapinb.com>"]

[tool.poetry.dependencies]
python = "^3.12"
wagtail = "^6.1"
python-decouple = "^3.8"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
mypy = "^1.10.0"
pytest = "^8.2.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "wagtail_site.settings.dev"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py"]
